collector-python/requirements.txt
requests
pika
python-dotenv
schedule
